Module mruby::def

Enums

EnclosingRubyScope
Typesafe wrapper for the RClass * of the enclosing scope for an mruby Module or Class.

Traits

ClassLike
ClassLike trait unifies class::Spec and module::Spec.

Define
Define trait allows a type to install classes, modules, and methods into an mruby interpreter.

Functions

rust_data_free⚠
A generic implementation of a Free function for mrb_values that store an owned copy of an Rc smart pointer.

Type Definitions

Free
Typedef for an mruby free function for an mrb_value with tt MRB_TT_DATA.

Method
Typedef for a method exposed in the mruby interpreter.