Module mruby::extn::core::regexp

ruby/spec compliant implementation of Regexp.

Each function on Regexp is implemented as its own module which contains the Args struct for invoking the function.

Modules

case_compare
Regexp#===

casefold
Regexp#casefold?

enc
Parse encoding parameter to Regexp#initialize and Regexp::compile.

eql
Regexp#eql? and Regexp#==

escape
Regexp::escape and Regexp::quote

fixed_encoding
Regexp#fixed_encoding?

hash
Regexp#hash

initialize
Regexp::new and Regexp::compile

inspect
Regexp#inspect

match_
Regexp#match

match_operator
Regexp#=~

match_q
Regexp#match?

named_captures
Regexp#named_captures

names
Regexp#names

options
Regexp#options

opts
Parse options parameter to Regexp#initialize and Regexp::compile.

source
Regexp#source

syntax
Helpers for parsing Regexp patterns.

to_s
Regexp#to_s

union
Regexp::union

Structs

Regexp

Functions

init