Struct mruby::value::Value

pub struct Value { /* fields omitted */ }

Wrapper around a sys::mrb_value.

Methods

impl Value

pub fn new(interp: &Mrb, value: mrb_value) -> Self

Construct a new Value from an interpreter and sys::mrb_value.

pub fn inner(&self) -> mrb_value

The sys::mrb_value that this Value wraps.

pub fn ruby_type(&self) -> Ruby

Return this values Rust-mapped type tag.

pub fn is_unreachable(&self) -> bool

Some type tags like MRB_TT_UNDEF are internal to the mruby VM and manipulating them with the sys API is unspecified and may result in a segfault.

After extracting a sys::mrb_value from the interpreter, check to see if the value is unreachable and propagate an MrbError::UnreachableValue error.

See: https://github.com/mruby/mruby/issues/4460

pub fn protect(&self)

Prevent this value from being garbage collected.

Calls sys::mrb_gc_protect on this value which adds it to the GC arena. This object will remain in the arena until ArenaIndex::restore restores the arena to an index before this call to protect.

pub fn is_dead(&self) -> bool

Return whether this object is unreachable by any GC roots.

pub fn to_s(&self) -> String

Call #to_s on this Value.

This function can never fail.

pub fn to_s_debug(&self) -> String

Generate a debug representation of self.

Format:

"#{self.class.name}<#{self.inspect}>"

This function can never fail.

pub fn inspect(&self) -> String

Call #inspect on this Value.

This function can never fail.

pub fn try_into<T>(self) -> Result<T, MrbError> where
T: TryFromMrb<Self, From = Ruby, To = Rust>,

Consume self and try to convert self to type T.

If you do not want to consume this Value, use Value::itself.

pub fn itself<T>(&self) -> Result<T, MrbError> where
T: TryFromMrb<Self, From = Ruby, To = Rust>,

Call #itself on this Value and try to convert the result to type T.

If you want to consume this Value, use Value::try_into.

pub fn freeze(self) -> Result<Self, MrbError>

Call #freeze on this Value and consume self.